% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow.R
\name{fetch_data}
\alias{fetch_data}
\title{Fetch Data}
\usage{
fetch_data(
  query,
  jdbc_Driver = jdbcDriver,
  path_env = "/home/rstudio/project/.env",
  snowflake_username = Sys.getenv("SNOWFLAKE_USER"),
  snowflake_password = Sys.getenv("SNOWFLAKE_PASSWORD"),
  warehouse = "USER_ADHOC",
  role = "BUSINESS_ANALYTICS",
  database = "GOLD",
  schema = "XDMADM"
)
}
\arguments{
\item{query}{(str) The SQL query to execute}

\item{jdbc_Driver}{The driver from the SnowflakeConnector}

\item{path_env}{(str) Path to .env file with credentials.  Default set in dockercontainer}

\item{snowflake_username}{(str) Username, default from .env}

\item{snowflake_password}{(str) password, default from .env}

\item{warehouse}{(str) warehouse, default "USER_ADHOC"}

\item{role}{(str) role, default "BUSINESS_ANALYTICS}

\item{database}{(str) database, default "GOLD"}

\item{schema}{(str) schema, defaul "XDMADM"}
}
\value{
Data Frame
}
\description{
This function uses the SnowflakeConnector to pull a data set.
}
\examples{
driver <- SnowflakeConnector()
q <- "SELECT * FROM GOLD.MRKT_FRCST.STG_WEATHER LIMIT 5"
fetch_data(query = q)
}
