% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow.R
\name{SnowflakeConnector}
\alias{SnowflakeConnector}
\title{Snowflake Connector}
\usage{
SnowflakeConnector(
  path_jar = "/home/rstudio/project/snowflake-jdbc-3.11.1.jar"
)
}
\arguments{
\item{path_jar}{(str) Path to the snowflake .jar file.  Default set in docker container}
}
\value{
The JDBC driver
}
\description{
This function sets up a connection with Snowflake
}
