% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{Test Function}
\usage{
test(test_word = "")
}
\arguments{
\item{test_word}{(str) Add to the test statement}
}
\value{
A str that says it works and the extra
}
\description{
This is a simple "hello world" type of function
}
\examples{
test("Whoohooo")
}
